requirements.txt
defusedxml==0.5.0
Django==1.10.6
django-actionlog==0.9.12
django-bootstrap-form==3.2.1
django-request==1.5.1
fluent-logger==0.4.6
igo-python==0.9.9
msgpack-python==0.4.8
numpy==1.12.0
oauthlib==2.0.1
pandas==0.19.2
psycopg2==2.7
PyJWT==1.4.2
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
